+++ examples/scripts/getting_started/maximum_a_posteriori.py
@@ -56,9 +56,8 @@ parameter_values.update(
 simulator = pybop.pybamm.Simulator(
     model, parameter_values=parameter_values, protocol=dataset
 )
-likelihood = pybop.GaussianLogLikelihood(dataset)
-posterior = pybop.LogPosterior(likelihood)
-problem = pybop.Problem(simulator, posterior)
+cost = pybop.GaussianLogLikelihood(dataset)
+log_pdf = pybop.LogPosterior(simulator, cost)
 
 # Set up the optimiser
 options = pybop.PintsOptions(
@@ -67,13 +66,13 @@ options = pybop.PintsOptions(
     min_iterations=20,
     max_iterations=50,
 )
-optim = pybop.XNES(problem, options=options)
+optim = pybop.XNES(log_pdf, options=options)
 
 # Run the optimisation
 result = optim.run()
 
 # Plot the timeseries output
-pybop.plot.problem(problem, inputs=result.best_inputs, title="Optimised Comparison")
+pybop.plot.problem(log_pdf, inputs=result.best_inputs, title="Optimised Comparison")
 
 # Plot the optimisation result
 result.plot_convergence()