+++ tests/unit/test_evaluation.py
@@ -69,7 +69,6 @@ class TestEvaluation:
             pybop.SumSquaredError,
             pybop.Minkowski,
             pybop.SumOfPower,
-            pybop.LogPosterior,
             pybop.GaussianLogLikelihood,
             pybop.GaussianLogLikelihoodKnownSigma,
         ]
@@ -78,10 +77,6 @@ class TestEvaluation:
         cost_class = request.param
         if cost_class is pybop.GaussianLogLikelihoodKnownSigma:
             cost = pybop.GaussianLogLikelihoodKnownSigma(dataset, sigma=0.002)
-        elif cost_class is pybop.LogPosterior:
-            cost = cost_class(
-                pybop.GaussianLogLikelihoodKnownSigma(dataset, sigma=0.002)
-            )
         else:
             cost = cost_class(dataset)
         return pybop.Problem(simulator, cost)